go.mod: Update github.com/containers/image/v5 to v5.10.4
containers/storage@v1.23.7 is imported by containers/image@v5.9.0, which
imports Microsoft/hcsshim@v0.8.9, which imports kubernetes/kubernetes,
which is not recommended for any consumer of Kubernetes and its
subproject libraries.

This was fixed in https://github.com/microsoft/hcsshim/pull/783, so
let's pull in a newer version of containers/image that reflects this
change.

Ultimately, this was _rediscovered_ when we replaced the verify
dependencies script in kubernetes/kubernetes with sigs.k8s.io/zeitgeist,
which imports k/release.

Signed-off-by: Stephen Augustus <foo@auggie.dev>