Refactor kubecross retrieval logic
We now introduce a new `kubecross` package which provides a new API for
retrieving the kubecross version for the default as well as a specific
branch. This API will be re-used to set the `KUBE_CROSS_VERSION_LATEST`
substitution, which is being used to build the `krel` binary before
building the actual release.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>