kube-cross: Adds mingw-w64 for Windows binary compilation
For compiling Windows binaries, we're typically using the docker image
dockcross/windows-static-x64, which is currently hosted on dockerhub.

Adding the mingw-w64 package into the kube-cross image will allow us to use
it for Windows binaries as well, so we won't have to rely on any docker image.
Because we won't have to pull another image, this will also speed up the building
process of some images (e.g: pause image)