Add CVE maps cloning and parsing to changelog
This commit adds new capability to the changelog package to clone CVE
data maps from a bucket and add the information to the changelog.

It adds a new function `changelog.CloneCVEData()` which clones the
CVE maps bucket directory and adds the temporary dir to the
changelog options.

When generating the release notes, changelog will pass the cloned dir
to the notes gatherer as a mapProvider.

It is important to note that the CVE data cloning and parsing only
has an effect when running for a patch release.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>