k8s-ci-builder: Match go version of the builder image with k/k@dev
This builder image is responsible for running compile-release-tools and
copying in tooling like `krel` into the final image.

AFAIK, the compiled RelEng tools have no requirement to exactly match
the Golang version of all active kubernetes/kubernetes release branches.

In instances where we make changes to releng tooling which are
backwards-incompatible, we have the possibility to cause image build
failures for other branch variants (as is currently happening).

Signed-off-by: Stephen Augustus <foo@auggie.dev>