pkg/cip: Revert os/io go1.16 changes
sigs.k8s.io/k8s-container-image-promoter is currently stuck on go1.15
due to a variety of bazel dependency updates that are not straightforward
to untangle.

Due to the incompatibility of the updated os/io packages with go1.15, we're
no longer able to update k/release in k-sigs/cip.

This commit reverts the go1.16 os/io usage to the go1.15-compatible ioutil
methods.

Signed-off-by: Stephen Augustus <foo@auggie.dev>