Generate and stage K8s source SBOM
This commits plugs the generation of the kubernetes source
code SBOM into the stage phase of the release process.

The source code SBOM is generated on lye once as all versions
in a release are cut at the same point in the commit history.

After generating the source SBOM, it is customized and tagged
with its own namespace before writing the SPDX file.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>