Fix bug downloading packages
This commit fixes a bug that was causing errors downloading go packages
by fixing two things:

1. When packages have no revision, ie are imported without specifying
a version, we now call the VCS at the latest revision, previously it
was calling go VCS with an empty string as revision.

2. We now only pull packages without version only when they are imported
that way. If we have a specific version we do not pull non versioned
packages anymore.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>