Add artifact verification step
This commit adds a new step to the stage phase: VerifyArtifacts()

This is the hook where we will perform as much verification of the
artifacts as possible before declaring stage as succcessfull and
releasing them.

This commit adds the hook, empty and regenerates the mocks. The next
ones will implement the first chunk of verification logic.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>