release.ArtifactChecker
This commit intoroduces the artifact checkes, a new object that
is intended to contain all the verifications that we want to
do o a staged build of Kubernetes.

This is the first commit which implements the verification of
tags in binaries. Once we start building more and more checks,
they will be added to this object and will be called during stage.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>