SPDX: Correctly form layer ref for single img ref
When saving a tar archive for an image reference, SPDX would
build an incomplete digests references for layers when the manifest
pointed to a single image (as opposed to an index). This bug caused
that generating SBOMs for single images stored in registries other
than docker hub would fail.

This commit fixes the references for laters in single images to
fix the bug.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>