Only warn on SLSA provenance errors
This commit modifies the call to CheckProvenance() during release. For
now, we will treat provenance errors as non-fatal and warn in the
release process log.

Once we finish testing and implementing SLSA1 support, this commit
should be reverted to reenable failing the build when the provenance
metadata does not match the staged artifacts.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>