spdx: Document outliner
This commit modifies the spdx package to support drawing an outline
of the elements contained in the document. The outline allows drawing
a tree-like structure to visualize the relationships expressed in the
SPDX code.

The first iteration of the outliner code supports defining the depth
of the recursive outline, expressing the tree nodes as ids or names
and other options which will be exposed through the main utility.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>