relnotes: Sleep when hitting the secondary GH limit
The last couple of releases have been hitting a new secondary rate limit
in the GitHub API. This PR makes some changes to play a bit more fair with
the API.

Now, the notes gatherer interprets the GitHub API responde. If the
secondary rate limit was hit, the functions that make a lot of calls
will now sleep for a minute+random before respawning.

Fixes 2225
Fixes 2302

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>