Rebase the main branch instead of merging it
This commit is a partial revert of
https://github.com/kubernetes/release/pull/2131

The reason why we are reverting this change is that merging on the main
(master) branch can cause the git mainline to end up in an undesired
state (i.e. it can become not straight). This can cause publishing-bot
to stop working or to miss commits around the release.

This issue can occur if there are merges between starting the stage
and running the release step. This often happens on the master
branch, especially when there's no code freeze in the effect
(e.g. for alpha/beta releases).

Signed-off-by: Marko Mudrinić <mudrinic.mare@gmail.com>