Support automatic krel fast forward
We now add an additional `--non-interactive` flag to `krel ff`, which
avoids requiring user interaction. If no `--branch` is provided, then
`krel ff` will no longer fail, but try to determine the latest release
branch. If this release branch is found, it checks if a fast forward is
required by testing if the latest final tag (like v1.23.0) exists on
the remote. If the tag exists, krel will will do nothing. Otherwise it
will do the fast forward.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>