kubepkg: loosen dependencies in packaging
At the moment, kubeadm had a hard dependency on the same version of
kubectl and kubelet, however, this is not a real dependency and it
can cause an unexpected upgrade of other packages if you are simply
installing kubeadm to prepare for an upgrade.

This loosens up the dependencies to maek sure that kubeadm only
depends on cri-tools which it really needs.  This will improve user
experience and unexpected upgrades.

It's expected that the user upgrades the packages for the other
systems correctly, and the upgrade documentation for kubeadm
explicitly asks them to upgrade those packages so they should be
done.