Fix changelog generation for custom forks
We now default to the base k/k repository when doing `krel stage` with a
custom Kubernetes fork. This allows us to generate the changelog from a
single source of truth, because we cannot know which modifications have
been done to the fork or if the tags/release branches are available
there.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>