distroless-iptables: fix `grep` command
The issue was that libaries from /usr/lib get symlinked from /lib, which
is just a symlink to /usr/lib in debian 12. To avoid that, we now always
refer to /usr/lib. To avoid wrong glibc versions we also have to use
distroless/base-debian12.

Bumped the version of the image to v0.3.1.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>