Add release-notes check subcommand
This commit adds a new "check" subcommand to the release-notes
utility.  This subcommand reads a PR from the github API and
check if it has a valid release note. If not it exits non-zero.

It is designed to easily add a PR check in CI to enforce authors
to add release notes.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>