[notes] Get the number of the origin PR for cherry-picks
Get the number of the origin PR for cherry-picks from the PR's
branch name instead of the commit message.

This way we get the same behavior, using the origin release note
if the current one is empty, for the cherry-picks created by the
`hack/..` script and those created by prow, regardless of the
merge strategy used in the project.