Do not exclude mapped notes
If a note PR get's modified to remove the release note, then the label
`release-note-none` will get applied to it. If a map for that note
exists, then it will excluded based on that. We now still provide the
mapped value to ensure a higher priority over the original PR.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>