Pin docker-ce-cli to version 24.0.x for API 1.43 compatibility
The latest docker-ce-cli package (API version 1.52) is incompatible
with the Docker-in-Docker setup in CI, which supports a maximum API
version of 1.43. This causes build failures with the error:
"client version 1.52 is too new. Maximum supported API version is 1.43"

Docker 24.0.x uses API version 1.43, making it compatible with the
current CI DinD environment.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>